+++ src/plugins/data_source/server/legacy/client_config.ts
@@ -15,13 +15,15 @@ import { DataSourcePluginConfigType } from '../../config';
 export function parseClientOptions(
   // TODO: will use client configs, that comes from a merge result of user config and default legacy client config,
   config: DataSourcePluginConfigType,
-  endpoint: string
+  endpoint: string,
+  registeredSchema: any[]
 ): ConfigOptions {
   const configOptions: ConfigOptions = {
     host: endpoint,
     ssl: {
       rejectUnauthorized: true,
     },
+    plugins: registeredSchema,
   };
 
   return configOptions;